implemented explore page

+++ src/lib/appwrite/api.ts
@@ -284,11 +284,53 @@ export async function deletePost(postId: string, imageId: string) {
 	if (!postId || !imageId) throw Error;
 
 	try {
-		await databases.deleteDocument(appwriteConfig.databaseId, appwriteConfig.postsCollectionId, postId);
+		const statusCode = await databases.deleteDocument(
+			appwriteConfig.databaseId,
+			appwriteConfig.postsCollectionId,
+			postId
+		);
+
+		if (!statusCode) throw Error;
 
 		await storage.deleteFile(appwriteConfig.storageId, imageId);
 
 		return { status: "ok" };
+	} catch (e) {
+		console.log(e);
+	}
+}
+
+export async function getInfinitePosts({ pageParam }: { pageParam: number }) {
+	const queries = [Query.orderDesc("$updatedAt"), Query.limit(10)];
+
+	if (pageParam) {
+		queries.push(Query.cursorAfter(pageParam.toString()));
+	}
+
+	try {
+		const posts = await databases.listDocuments(
+			appwriteConfig.databaseId,
+			appwriteConfig.postsCollectionId,
+			queries
+		);
+
+		if (!posts) throw Error;
+
+		return posts;
+	} catch (e) {
+		console.log(e);
+	}
+}
+
+export async function searchPosts(searchTerm: string) {
+	try {
+		const posts = await databases.listDocuments(appwriteConfig.databaseId, appwriteConfig.postsCollectionId, [
+			Query.search("caption", searchTerm),
+		]);
+
+		if (!posts) throw Error;
+
+		return posts;
 	} catch (e) {
 		console.log(e);
 	}
